move viewTypes to types/index
