+++ src/components/News/NewsThumb.vue
@@ -41,7 +41,7 @@
 <script lang="ts">
 import { defineComponent, PropType } from "@vue/composition-api";
 import Icon from "vue-awesome/components/Icon.vue";
-import { NewsPost } from "@/types/viewTypes";
+import { NewsPost } from "@/types";
 import AppUrls from "@/utils/urls";
 import useNews from "@/factories/useNews";
 import ProgressiveImage from "@/components/ProgressiveImage.vue";
