+++ src/components/Programs/schoolProgramsTabs.ts
@@ -4,7 +4,7 @@ import ProgressiveImage from "@/components/ProgressiveImage.vue"
 import DownloadLink from "@/components/DownloadLink.vue"
 import useSchoolPrograms from "@/factories/useSchoolPrograms"
 import useTabs from "@/utils/useTabs"
-import { SchoolProgram } from "@/types/viewTypes"
+import { SchoolProgram } from "@/types"
 
 const SchoolProgramsTabs = defineComponent({
   name: "SchoolProgramsTabs",
