+++ src/factories/usePrograms.ts
@@ -1,7 +1,7 @@
 import Vue from "vue"
 import VueCompositionApi, { computed } from "@vue/composition-api"
 import apiRoutes from "../../api/apiRoutes"
-import { Program } from "@/types/viewTypes"
+import { Program } from "@/types"
 import useAsyncData from "../utils/useAsyncData"
 import {
   WPResponseItem,
