+++ src/factories/useSponsors.ts
@@ -5,7 +5,7 @@ import useAsyncData from "@/utils/useAsyncData"
 import { WpResponseData, WPResponseItem } from "@/types/wordpressTypes"
 import { getCustomField, getWPTitle } from "@/utils/api"
 import { SponsorCategoryKeys, SponsorKeys } from "@/types/customFieldsKeysTypes"
-import { Sponsor, SponsorsCategory } from "@/types/viewTypes"
+import { Sponsor, SponsorsCategory } from "@/types"
 import { filterUndef, sortByOrder } from "@/utils/arrays"
 
 Vue.use(VueCompositionApi)
