+++ src/utils/arrays.ts
@@ -1,9 +1,4 @@
-import {
-  CarouselImage,
-  Document,
-  SponsorsCategory,
-  Sponsor,
-} from "@/types/viewTypes"
+import { CarouselImage, Document, SponsorsCategory, Sponsor } from "@/types"
 
 export function filterUndef<T>(ts: (T | undefined)[]): T[] {
   return ts.filter((t: T | undefined): t is T => !!t)
